# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Restaurante Piscina",
        "description": "API para gestión de restaurante con módulos de cocina, meseros, órdenes, facturación y más",
        "contact": {
            "name": "Soporte Técnico",
            "email": "soporte@restaurante-piscina.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/recipes": {
            "get": {
                "tags": [
                    "Recetas"
                ],
                "summary": "Listar todas las recetas",
                "operationId": "e6ebf01e62f938e7ce367c7b2fc94110",
                "responses": {
                    "200": {
                        "description": "Lista de recetas con ingredientes"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Recetas"
                ],
                "summary": "Crear nueva receta",
                "operationId": "5ed3009a1c8ce5b784cac78533de2c73",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "ingredients"
                                ],
                                "properties": {
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Pizza Margherita - Receta Base"
                                    },
                                    "ingredients": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "ingredient_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "description": "Cantidad en la unidad del ingrediente",
                                                    "type": "number",
                                                    "example": 0.2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Receta creada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/recipes/{id}": {
            "get": {
                "tags": [
                    "Recetas"
                ],
                "summary": "Obtener detalle de una receta",
                "operationId": "e59df102512a161f3e7be88aaa85a0e1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle de la receta con ingredientes"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Recetas"
                ],
                "summary": "Actualizar receta",
                "operationId": "573ece0098b0f511b333f1406621b99c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Receta actualizada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Recetas"
                ],
                "summary": "Eliminar receta",
                "operationId": "a527903d2d7195e7456416e97f55faf2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Receta eliminada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/recipes/product/{productId}": {
            "get": {
                "tags": [
                    "Recetas"
                ],
                "summary": "Obtener recetas de un producto específico",
                "operationId": "dc3c7efc00e7c0f7f272078256de7ef1",
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recetas del producto"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token JWT obtenido del login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}